+++ ldr_setup.h
@@ -1,4 +1,4 @@
 /* LDR Setup */
-int sensorPin = 0;
+byte sensorPin = 0;
 int ldr_int = 0;
 char ldr_str[4];
